Fixing cookies + User-Agent
The script stopped working. The issue was the lack of a 'user' cookie.
This cookie gets added by the request that posts the login data. This
response adds this cookie, and 302 redirects back to the front page.
This cookie doesn't seem to get added unless I change the User-Agent
script away from the default python-requests User-Agent string.

It now has a proper User-Agent string that even includes the URL of the
Github page. I'm unsure if the 'user' cookie is new, or if the
blacklisting of the python-requests default user agent is new.