Fixing date format issues with HTTP headers
Apparently sometimes the %B is not the correct format. As this happened
recently, maybe whatever is generating the Date: header server-side is
not abbreviating June to Jun (as per %B format) even though it uses
short forms for the other months. Another possibility is that they
recently changed the date format as delivered from the server.

Either way, this work-around will parse whichever format shows up.